Complete walking skeleton for command line app
"walking skeleton" is a concept from the book "Growing Object-Oriented
Software Guided by Tests":

    A “walking skeleton” is an implementation of the thinnest possible
    slice of real functionality that we can automatically build, deploy,
    and test end-to-end

I like starting projects like this for a couple reasons (which the book
explains very well):

- From this point on you can feel a certain sense of confidence that
  your future changes will be tested.
- Attempting to automate the build, deploy, test pipeline forces you to
  think about what other services this application talks to which in
  turn leads to you making sure that you have the necessary credentials
  to talk with said services.

For the command line portion of this application there is not much to do
since:
- There is nothing to build or deploy since this is a command line
  application (although I suppose I could build binaries and push them
  to github like https://github.com/schollz/croc but that seems a little
  overkill for this project)
- It does not talk to other services

But! We still got a useful thing out of this exercise namely an
automated testing infrastructure.